Fix increment/decrement applying to any expression
Since this alters the value, it only makes sense on identifiers (i.e.
variables). If the old parser allowed that, we'll have to make another
breaking change.

Was causing ambiguities when parsing the following wizard:

  ++a
  --a
  a++
  a--

This would get parsed as an expr containing an expr containig the ++,
and so on - when it should just be four exprs, one for each
increment/decrement.